+++ src/app/course/_test/course-events-snippet/course-events-snippet.component.spec.ts
@@ -3,7 +3,7 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {CourseEventsSnippetComponent} from '../../course-events-snippet/course-events-snippet.component';
 import {TestModule} from '@test/test.module';
 import {CourseEventService} from "@app/course/_services/course-event.service";
-import {CourseEventServiceMock} from "@app/problems/_test/course-event.service.mock";
+import {CourseEventServiceMock} from "@app/problems/_test/_services/course-event.service.mock";
 import {MOCK_EVENT_TYPES, MOCK_USER_STUDENT, MOCK_USER_TEACHER} from "@app/course/_test/mock";
 import {MOCK_COURSE, MOCK_COURSE_EVENT} from "@app/problems/_test/mock";
 import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
