+++ src/app/course/_test/course-question-snippet/course-question-snippet.component.spec.ts
@@ -5,9 +5,9 @@ import {TestModule} from '@test/test.module';
 import {CourseService} from "@app/course/_services/course.service";
 import {CourseServiceMock} from "@test/course.service.mock";
 import {CourseEventService} from "@app/course/_services/course-event.service";
-import {CourseEventServiceMock} from "@app/problems/_test/course-event.service.mock";
+import {CourseEventServiceMock} from "@app/problems/_test/_services/course-event.service.mock";
 import {UqjService} from "@app/problems/_services/uqj.service";
-import {UqjServiceMock} from "@app/problems/_test/uqj.service.mock";
+import {UqjServiceMock} from "@app/problems/_test/_services/uqj.service.mock";
 import {ActivatedRoute, convertToParamMap, RouterModule} from "@angular/router";
 import {MOCK_USER_TEACHER, MOCK_UQJ, MOCK_UQJ2} from "@app/course/_test/mock";
 import {RouterTestingModule} from "@angular/router/testing";
