+++ src/app/course/_test/token-use-snippet/token-use-snippet.component.spec.ts
@@ -5,7 +5,7 @@ import {TestModule} from '@test/test.module';
 import {MOCK_COURSE_REGISTRATION, MOCK_TOKEN_USE1} from "@app/course/_test/mock";
 import {ActivatedRoute} from "@angular/router";
 import {TokenUseService} from "@app/course/_services/token-use.service";
-import {TokenUseServiceMock} from "@app/course/_test/token-use.service.mock";
+import {TokenUseServiceMock} from "@app/course/_test/_services/token-use.service.mock";
 import {CourseModule} from "@app/course/course.module";
 
 describe('TokenUseSnippetComponent', () => {
