+++ src/app/problems/_test/problem-edit/parsons-edit-snippet.component.spec.ts
@@ -11,8 +11,8 @@ import {CategoryService} from "@app/_services/api/category.service";
 import {CategoryServiceMock} from "@test/category.service.mock";
 import {CourseServiceMock} from "@test/course.service.mock";
 import {QuestionService} from "@app/problems/_services/question.service";
-import {QuestionServiceMock} from "@app/problems/_test/question.service.mock";
-import {CourseEventServiceMock} from "@app/problems/_test/course-event.service.mock";
+import {QuestionServiceMock} from "@app/problems/_test/_services/question.service.mock";
+import {CourseEventServiceMock} from "@app/problems/_test/_services/course-event.service.mock";
 import {CourseService} from "@app/course/_services/course.service";
 import {CourseEventService} from "@app/course/_services/course-event.service";
 import {Router} from "@angular/router";
