+++ src/app/problems/_test/problem-set/problem-set.component.spec.ts
@@ -5,12 +5,12 @@ import {TestModule} from '@test/test.module';
 import {CategoryService} from "@app/_services/api/category.service";
 import {CategoryServiceMock} from "@test/category.service.mock";
 import {DifficultyService} from "@app/problems/_services/difficulty.service";
-import {DifficultyServiceMock} from "@app/problems/_test/difficulty.service.mock";
+import {DifficultyServiceMock} from "@app/problems/_test/_services/difficulty.service.mock";
 import {ReactiveFormsModule} from "@angular/forms";
 import {MatPaginatorModule} from "@angular/material/paginator";
 import {MatTableModule} from "@angular/material/table";
 import {QuestionService} from "@app/problems/_services/question.service";
-import {QuestionServiceMock} from "@app/problems/_test/question.service.mock";
+import {QuestionServiceMock} from "@app/problems/_test/_services/question.service.mock";
 import {FontAwesomeModule} from "@fortawesome/angular-fontawesome";
 import {AppRoutingModule} from "@app/app-routing.module";
 import {MOCK_DIFFICULTIES} from "@app/problems/_test/mock";
