+++ src/app/problems/_test/problem-view/java-view-snippet.component.spec.ts
@@ -4,7 +4,7 @@ import {JavaViewSnippetComponent} from '../../problem-view/java-view-snippet/jav
 import {TestModule} from '@test/test.module';
 import {MOCK_UQJ_3, MOCK_UQJS} from '@app/problems/_test/mock';
 import {SubmissionService} from "@app/problems/_services/submission.service";
-import {SubmissionServiceMock} from "@app/problems/_test/submission.service.mock";
+import {SubmissionServiceMock} from "@app/problems/_test/_services/submission.service.mock";
 
 describe('JavaViewSnippetComponent', () => {
     let component: JavaViewSnippetComponent;
